+++ .github/workflows/python-app.yml
@@ -47,3 +47,7 @@ jobs:
       - name: Test with pytest
         run: |
           pytest --cov=tests --cov-report=xml
+
+      - name: Run main
+        run: |
+          python main.py
