+++ Dockerfile
@@ -29,4 +29,5 @@ RUN pip install --no-cache-dir -r requirements.txt
 EXPOSE 5002
 
 # Run app.py when the container launches
-CMD ["python3", "main.py"]
+CMD ["cd", "api"]
+CMD ["python3", "api/main.py"]
