+++ requirements.txt
@@ -1,2 +1,3 @@
 python-socketio==5.4.0
 aiohttp[speedups]
+pytest==7.4.3
